\name{bwsurvival-internal}
\alias{categ}
\alias{count}
\alias{pos}
\alias{valkm}
\alias{table}

\title{Internal bwsurvival functions}
\description{
Auxiliary functions used in the package
Author: C. Serrat and G. Gómez}
\usage{
 categ(data,breaks)
 count(data,breaks)
 pos(data, seq)
 valkm(km, x = 0.)
 table(bwsurvival,timesofinterest)
 }
\details{These functions are not to be called by the user.}
\keyword{internal}